api: change --disable_sudo -> --sudo
Align with options in train. This also prevents unexpected behavior where sudo is used, without being requested. Example:

```
command('uname -a')
```

This command will not run if no sudo information is given and e.g. a sudo password is configured on the node. The user can activate sudo usage with `--sudo` now.