Adopting PR 1570 - fixing processes resource doc
As originally proposed in #1570, the documentation of
the `processes` resource could be clearer in its ability
to accept a regular expression. The original doc-only PR
was missing a DCO sign-off and had one required change
before it was able to be merged. The PR unfortunately
became stale.

Adopting the PR and merging under the obvious-fix /
doc-only-fix rule.

Signed-off-by: Adam Leff <adam@leff.co>