[www] Update www Gemfile.lock
The Gemfile.lock in the www directory needs to be updated since
the last InSpec release. We should probably do this as part of
the normal InSpec release process.

You need to `bundle install` before releasing the web stuff,
and the www rake tasks will fail if your git repo is unclean.

Signed-off-by: Adam Leff <adam@leff.co>