feat: Adds port/ protocol support for host
This update extends the host resource to use correct
port and protocol for the redhat and debian families.
It uses a feature from bash and the timeout command
from the coreutils packages.

When port is specified but protocol is omitted, it
assumes tcp is what was intended. When neither port
nor protcol is specified it falls back to ping the
host.