Remove controls from JSON formatter that have no results
When profile dependencies/inheritance is used, controls from
depended-on profiles are tracked in the "including profile" run
context as well as the "depended-on profile." When you ask the
"including profile" for its info, it returns its controls as well
as controls from the "depended-on profile."

In the JSON formatter, this causes these depended-on controls to
be output twice, once with results and once without. Systems that
consume this JSON can incorrectly show these controls as skipped.

Until we dig into how best to separate profile run contexts, I'm
proposing removing any controls from profile output that have no
results. This will better align it with the CLI formatter and
avoid downstream systems from seeing the same control twice.

Signed-off-by: Adam Leff <adam@leff.co>