Add warning for potential empty password option
If a user supplies --password on the command line but fails to provide a value,
Thor will set the value of the `password` option to the word "password". Users
may expect to be prompted for a password if leaving this blank, but InSpec will
happily try to log in to the target system with the password of "password".

Unfortunately, there is no way to alter Thor's behavior for this, and users may
actually have a password of "password".

Proposing we add a warning if the password is "password" in an attempt to catch
those users that accidentally provide an empty password. It may also serve as a
warning to users that legitimately have a password of "password" they should
perhaps use a different phrase. :)

Signed-off-by: Adam Leff <adam@leff.co>