Remove www Gemfile.lock
No reason to keep the Gemfile.lock in the www directory under source
control. It adds extra steps to the www release process after an InSpec
release because the release process will fail of the git working tree
is not clean (and since the www Rake tasks rely on InSpec stuff, the
Gemfile.lock will be perpetually out of date).

Signed-off-by: Adam Leff <adam@leff.co>