[win-perf] security policies
this change executes and memoizes the results of the `secedit /export`
command once. previously, we'd issue three powershell commands for
each `security_policy` block in a profile. after this change, we issue
the same three powershell commands, but only once. This reduces the
number of powershell commands executed from 3N to 3.

Signed-off-by: Stephen Delano <stephen@chef.io>