Updates README to use function instead of alias
I really liked the inclusion of the alias for the docker run version
of installing inspec. The instructions tried to make it clear that
when you create the alias that it was for only your current directory
but it sounded more like your current working directory wherever you
were so I thought that it would be better to actually make that work.

By changing it to a function you now can delay the processing of the
$(pwd) subshell so that it will mount your current working directory
when you run it. This makes it really easy to use the InSpec CLI
and the installation this way is on par with the other ways.

Signed-off-by: Franklin Webber <franklin@chef.io>