Allow apache_conf include files to be empty
In the `read_file` method we can call `read_file_content` with the second
argument `true` to avoid skipping on an empty file.

We will still skip the control if the main configuration file is empty
as there is still an explicit call to `read_file_content` without this
argument in the `read_content` method.

Signed-off-by: Kris Shannon <k.shannon@amaze.com.au>