adding a property of `file_count` to `directory` resource
This is helpful to support testing if a directory is empty, not empty or even in some cases matching a specific number.

While refactoring some chef changes I wanted to make sure that a particular directory includes no files as the client should not need them. They are needed on the server and the directory is automatically created on both client and server.

Signed-off-by: Ben Abrams <me@benabrams.it>