Modify URL fetcher to accept URI
When https://github.com/inspec/inspec/pull/3562 was merged it added a
method to parse a URI passed as a URL. However, some users (at least the
Audit cookbook) were passing a URI and not a URL so there is no need to
parse it.

This adds code to use the URI if a URI was passed.

Many thanks to @teknofire for discovering this and submitting a very
detailed issue.

Signed-off-by: Jerry Aldrich <jerryaldrichiii@gmail.com>