Remove Habitat package run time dependencies from Habitat plan (#3645)
* Remove Hab pkg deps in favor of OS binaries

This removes the runtime dependencies on Hab pkgs and instead modifies
the `PATH` environment variable to use the OS binaries where the InSpec
Habitat package is installed.

It should be noted that this is counter to what Habitat intends in most
cases. In general, it is preferable to use only Habitat packages as
runtime dependencies to get all the benefits that Habitat provides.

We elected not to do this for the InSpec Habitat package since the list
of binaries that would need to be installed to support all InSpec
resources would be prohibitively expensive (both in disk space and
network requirements). If you wish to use Habitat packaged binaries with
this package you can use `hab pkg install origin/my-binary --binlink`.

Signed-off-by: Jerry Aldrich <jerryaldrichiii@gmail.com>

* Modify Habitat install example to use `--binlink`

Signed-off-by: Jerry Aldrich <jerryaldrichiii@gmail.com>

* Add `core/git` runtime dep (used for fetching)

Signed-off-by: Jerry Aldrich <jerryaldrichiii@gmail.com>