Fix input hash processing on inheritance
When attempting to use an inherited profile, I was getting an error
indicating that `to_hash` could not be performed on an Inspec::Input
object.

This patch only calls the to_hash method if it is available on the
corresponding object.

No differences have been found in the output or processing.

Signed-off-by: Trevor Vaughan <tvaughan@onyxpoint.com>