Added test:knapsack task (draft).
This requires an external artifact to be generated and stored in
BK (which we're close to having), but once we have that this runs much
better than test:parallel does because it evens out all the work
across the workers by file run time.

Until we have artifacts cached, this is useless.

Signed-off-by: Ryan Davis <zenspider@chef.io>