Implement optional InSpecStyle in `inspec check`
With the `--inspecstyle` flag after `inspec check`, this will execute
a full run of the `inspecstyle` gem against the target file or
directory. As the aim is to have this default in InSpec 5.

It also includes additional information including how to get involved in
the discussion around future InSpecStyle rules; and that there are other
InSpecStyle tools available.