Modified windows_feature to indicate enabled rather than just available
There are three conditions a windows feature may be in as represented by DISM:

1. Available in the OS and enabled
       dism returns 0 and dism output of state = 'Enabled'
2. Available in the OS but disabled
       dism returns 0 and dism output of state != 'Enabled'
3. Not available  
       dism returns != 0 

This change, in effect, modifies the meaning of installed == true to represent that the feature is both available and enabled. 
installed == false will now indicate the feature is either unavailable or disabled.
The prior implementation indicated the feature was available, but did not recognize the it might be disabled. The most logical use for an InSpec evaluation of a Windows Feature seems to be determining if the feature is enabled or disabled rather than simply available. This updated implementation better represents that goal. 

----------

Developer's Certificate of Origin 1.1

By making a contribution to this project, I certify that:

(a) The contribution was created in whole or in part by me and I
    have the right to submit it under the open source license
    indicated in the file; or

(b) The contribution is based upon previous work that, to the
    best of my knowledge, is covered under an appropriate open
    source license and I have the right under that license to
    submit that work with modifications, whether created in whole
    or in part by me, under the same open source license (unless
    I am permitted to submit under a different license), as
    Indicated in the file; or

(c) The contribution was provided directly to me by some other
    person who certified (a), (b) or (c) and I have not modified
    it.

(d) I understand and agree that this project and the contribution
    are public and that a record of the contribution (including
    all personal information I submit with it, including my
    sign-off) is maintained indefinitely and may be redistributed
    consistent with this project or the open source license(s)
    involved.

Signed-off-by: Jeff Dean <jwdean@scriptpro.com>