Enable repeatable builds by bundling Gemfile.lock
This way changes to omnibus-software or omnibus don't change the builds. This is important for us, but also for cinc. It also makes sure that we're tracking bumps to the omnibus-software in the InSpec changelog so we can build release notes that reflect key dep changes like openssl. We recently added the same thing to Infra Server to better track deps there and we've done this in infra client for many years.

Signed-off-by: Tim Smith <tsmith@chef.io>